chore(cypress-ete): split `registration_2.7.cy.ts` into individual tests
